requirements.txt
fastapi
uvicorn
psycopg2-binary
redis
sqlalchemy
python-dotenv
celery
requests
python-jose
python-multipart
passlib
bcrypt
